% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotmm.R
\name{barplotmm}
\alias{barplotmm}
\title{barplotmm}
\usage{
library(ferroviz)

data(resultmm)

data(ferrdbmm)

barplotmm(resultmm,ferrdbmm,fc=0.25,size=16)
}
\arguments{
\item{resultmm}{output results of murine differential expressed genes found with limma}

\item{ferrdbmm}{Mus musculus ferrDB V2 database}

\item{fc}{threshold log2 fold change in absolute value}

\item{size}{size of the text in the figure}
}
\description{
the "barplotmm" for Mus musculus specie count differential expressed
genes regulated and implicated in ferroptosis (drivers and suppressors) from a limma
two groups analysis
}
\examples{
library(ferroviz)
data(resultmm)
data(ferrdbmm)
barplotmm(resultmm,ferrdbmm,fc=0.25,size=16)
}
