% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsvol.R
\name{hsvol}
\alias{hsvol}
\title{hsvol}
\usage{
library(ferroviz)

data(resulths)

data(ferrdbhs)

hsvol(resulths,ferrdbhs,color="grey",size=16,x=0.25,label=1.5)
}
\arguments{
\item{resulths}{output results of human differential expressed genes found with limma}

\item{ferrdbhs}{Homo sapiens ferrDB V2 database}

\item{color}{color of gene dots unclassed in ferrdb}

\item{size}{size of the text in the figure}

\item{x}{x coordinates of logFC for vertical lines}

\item{label}{size of text gene labels}
}
\description{
the "hsvol" function allows to draw volcanoplot of ferroptosis genes
(red: drivers) (blue: suppressors), from a limma output between 2 experimental
groups (Homo sapiens).
}
\examples{
library(ferroviz)
data(resulths)
data(ferrdbhs)
hsvol(resulths,ferrdbhs,color="grey",size=16,x=0.25,label=1.5)
}
