% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listmm.R
\name{listmm}
\alias{listmm}
\title{listmm}
\usage{
library(ferroviz)

data(resultmm)

data(ferrdbmm)

df<-listmm(resultmm,ferrdbmm,fc=0.10)

df
}
\arguments{
\item{resultmm}{output results of human differential expressed genes found with limma}

\item{ferrdbmm}{Homo sapiens ferrDB V2 database}

\item{fc}{threshold log2 fold change in absolute value}
}
\description{
the "litmm" for Mus musculus specie return annotated ferroptosis gene list
from limma output result between two groups of transcriptome experiments
}
\examples{
library(ferroviz)
data(resultmm)
data(ferrdbmm)
df<-listmm(resultmm,ferrdbmm,fc=0.10)
df
}
