% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmvol.R
\name{mmvol}
\alias{mmvol}
\title{mmvol}
\usage{
library(ferroviz)

data(resultmm)

data(ferrdbmm)

mmvol(resultmm,ferrdbmm,color="grey",size=16,x=0.25,label=1.5)
}
\arguments{
\item{color}{color of gene dots unclassed in ferrdb}

\item{size}{size of the text in the figure}

\item{x}{x coordinates of logFC for vertical lines}

\item{label}{size of text gene labels}

\item{resulths}{output results of human differential expressed genes found with limma}

\item{ferrdbhs}{Homo sapiens ferrDB V2 database}
}
\description{
the "mmvol" function allows to draw volcanoplot of ferroptosis genes
(red: drivers) (blue: suppressors), from a limma output between 2 experimental
groups (Mus musculus).
}
\examples{
library(ferroviz)
data(resultmm)
data(ferrdbmm)
mmvol(resultmm,ferrdbmm,color="grey",size=16,x=0.25,label=1.5)
}
